modify part1_2
